[#IOPID-1797] Refactor redis repo type
